feat: update signin by phonenumbers

+++ src/main/java/com/mfa/report/endpoint/rest/model/DTO/PerfRealizationDTO.java
@@ -12,7 +12,7 @@ import lombok.NoArgsConstructor;
 @Data
 public class PerfRealizationDTO {
   private String id;
-  private Long indicators;
+  private Double indicators;
   private String realization;
   private String activityId;
   private String realizationType;
